Refactor btc/bch using a common manager class
