+++ rotkehlchen/chain/bitcoin/btc/constants.py
